+++ RDBqc/R/MEDBS_check_duplicates.r
@@ -1,12 +1,12 @@
 #' Check for duplicated data rows in landings, discards and catch tables
 #'
 #' @param data data frame containing either landings, discards or catch data
-#' @param type string vector indicating the type of table to be checked. "l" for landing; "d" for discards; "c" for catch table.
+#' @param type string vector indicating the type of table to be checked. "l" for landing; "d" for discards; "c" for catch table; "gp" for growth parameters table; "alk" for age-length keys table; "ma" for maturity at age table; "ml" for maturity at length table; "sra" for sex ratio at age table; "srl" for maturity at length table..
 #' @param SP species code
 #' @param MS member state code
 #' @param GSA GSA code (Geographical sub-area)
 #' @param verbose boolean. If TRUE messages are returned
-#' @description The function checks the presence of duplicated rows in landings, discards or catch data.
+#' @description The function checks the presence of duplicated rows in MED & BS tables data.
 #' @return The function returns a data frame containing the duplicated rows to be likely deleted from the data.
 #' @examples
 #' MEDBS_check_duplicates(
@@ -27,11 +27,11 @@
 
 MEDBS_check_duplicates <- function(data, type = "l", SP, MS, GSA, verbose = TRUE) {
   if (FALSE) {
-    data <- Catch_tab_example
-    type <- "c"
+    data <- GP_tab_example
+    type <- "gp"
     MS <- "ITA"
-    GSA <- "GSA 9"
-    SP <- "DPS"
+    GSA <- "GSA 18"
+    SP <- "MUT"
     verbose <- TRUE
 
     MEDBS_check_duplicates(data = data, type = "l", MS = "ITA", GSA = "GSA 18", SP = "NEP", verbose = FALSE)
@@ -57,11 +57,38 @@ MEDBS_check_duplicates <- function(data, type = "l", SP, MS, GSA, verbose = TRUE
   if (type %in% c("l", "d")) {
     duplicate_rows <- table(duplicated(data[, c(2:13)])) # FALSE means no duplicated records are present
     row_to_del <- which(duplicated(data[, c(2:13)]))
-    dupl <- data[row_to_del, c(1:12)]
+    if (length(row_to_del)!=0){
+      dupl <- data[row_to_del, c(1:13)]    }
   } else if (type == "c") {
     duplicate_rows <- table(duplicated(data[, c(2:24)])) # FALSE means no duplicated records are present
     row_to_del <- which(duplicated(data[, c(2:24)]))
-    dupl <- data[row_to_del, c(1:24)]
+    if (length(row_to_del)!=0){
+      dupl <- data[row_to_del, c(1:24)]
+    }
+  } else if (type=="gp") {
+    duplicate_rows <- table(duplicated(data[, c(1:6)]))
+    row_to_del <- which(duplicated(data[, c(1:6)]))
+    if (length(row_to_del)!=0){
+      dupl <- data[row_to_del, c(1:6)]
+    }
+  } else if (type=="alk") {
+    duplicate_rows <- table(duplicated(data[, c(1:7,12)]))
+    row_to_del <- which(duplicated(data[, c(1:7,12)]))
+    if (length(row_to_del)!=0){
+      dupl <- data[row_to_del, c(1:7,12)]
+    }
+  } else if (type=="ma" | type=="ml") {
+    duplicate_rows <- table(duplicated(data[, c(1:7)]))
+    row_to_del <- which(duplicated(data[, c(1:7)]))
+    if (length(row_to_del)!=0){
+      dupl <- data[row_to_del, c(1:7)]
+    }
+  }  else if (type=="sra" | type=="srl") {
+    duplicate_rows <- table(duplicated(data[, c(1:6)]))
+    row_to_del <- which(duplicated(data[, c(1:6)]))
+    if (length(row_to_del)!=0){
+      dupl <- data[row_to_del, c(1:6)]
+    }
   }
 
   if (verbose) {