+++ RDBqc/NEWS.md
@@ -1,5 +1,6 @@
 # RDBqc 0.0.17
  * improving results of MEDBS functions (ALK, LFD, yr_missing_length) 
+ * expansion of MEDBS_check_duplicates function to all MED&BS tables of biological parameters (GP, ALK, MA, ML, SRA, SRL)
 
 # RDBqc 0.0.16
  * update compatibility with R version 4.3